+++ src/app.rs
@@ -3,6 +3,8 @@ use crate::components::schedule::ScheduleState;
 use crate::components::standings::StandingsState;
 use crate::components::stats::StatsState;
 use crate::config::{CONFIG_LOCATION, generate_config_file, load_config_file};
+use chrono::{NaiveDate, ParseError, TimeZone, Utc};
+use chrono_tz::{OffsetName, Tz};
 use crossbeam_channel::{Receiver, Sender, bounded};
 use mlb_api::client::{MLBApi, MLBApiBuilder};
 use mlb_api::live::LiveResponse;
@@ -62,10 +64,12 @@ pub struct AppState {
     pub stats: StatsState,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Default, Clone)]
 pub struct AppSettings {
     pub favorite_team: Option<String>,
     pub full_screen: bool,
+    pub timezone: Tz,
+    pub timezone_abbreviation: String,
 }
 
 pub struct App {
@@ -81,10 +85,7 @@ impl App {
     pub fn new() -> Self {
         let mut app = Self {
             state: AppState::default(),
-            settings: AppSettings {
-                favorite_team: None,
-                full_screen: false,
-            },
+            settings: AppSettings::default(),
             client: MLBApiBuilder::default().build().unwrap(),
             redraw_channel: bounded(1),
             update_channel: bounded(1),
@@ -103,10 +104,24 @@ impl App {
             }
             let config = load_config_file(&path)?;
             self.settings.favorite_team = config.validate_favorite_team();
+            self.settings.timezone = config.validate_timezone();
+            self.settings.timezone_abbreviation = self.get_timezone_abbreviation();
         } else {
             eprintln!("could not find config file");
         };
         Ok(())
+    }
+
+    /// Get the abbreviated name of the configured timezone, (e.g. "PST" or "PDT")
+    fn get_timezone_abbreviation(&self) -> String {
+        // Get the timezone offset for the given datetime
+        let now = Utc::now().with_timezone(&self.settings.timezone);
+        let offset = self
+            .settings
+            .timezone
+            .offset_from_utc_datetime(&now.naive_utc());
+
+        offset.abbreviation().unwrap_or("~~").to_string()
     }
 
     pub fn update_schedule(&mut self, schedule: &ScheduleResponse) {
@@ -125,6 +140,35 @@ impl App {
         }
     }
 
+    pub fn try_update_date_from_input(&mut self) -> Result<(), ParseError> {
+        let valid_date = self
+            .state
+            .date_input
+            .validate_input(self.settings.timezone)?;
+
+        // current tab is date picker, so use previous tab to update correct date
+        match self.state.previous_tab {
+            MenuItem::Scoreboard => self.state.schedule.set_date_from_valid_input(valid_date),
+            MenuItem::Standings => self.state.standings.set_date_from_valid_input(valid_date),
+            MenuItem::Stats => self.state.stats.set_date_from_valid_input(valid_date),
+            _ => (),
+        }
+        Ok(())
+    }
+
+    pub fn move_date_selector_by_arrow(&mut self, right_arrow: bool) {
+        let date = match self.state.previous_tab {
+            MenuItem::Scoreboard => Some(self.state.schedule.set_date_with_arrows(right_arrow)),
+            MenuItem::Standings => Some(self.state.standings.set_date_with_arrows(right_arrow)),
+            MenuItem::Stats => Some(self.state.stats.set_date_with_arrows(right_arrow)),
+            _ => None,
+        };
+        self.state.date_input.text.clear();
+        if let Some(date) = date {
+            self.state.date_input.text.push_str(&date.to_string());
+        }
+    }
+
     pub fn exit_help(&mut self) {
         if self.state.active_tab == MenuItem::Help {
             self.state.active_tab = self.state.previous_tab;
@@ -140,6 +184,18 @@ impl App {
 
     pub fn toggle_full_screen(&mut self) {
         self.settings.full_screen = !self.settings.full_screen;
+    }
+}
+
+impl DateInput {
+    pub fn validate_input(&mut self, tz: Tz) -> Result<NaiveDate, ParseError> {
+        let input: String = self.text.drain(..).collect();
+        let date = match input.as_str() {
+            "today" => Ok(Utc::now().with_timezone(&tz).date_naive()),
+            _ => NaiveDate::parse_from_str(input.as_str(), "%Y-%m-%d"),
+        };
+        self.is_valid = date.is_ok();
+        date
     }
 }